format the list so that it displays the pokemon and names flexibly
as a first pass on styling, the pokemons are all displayed in a
flexible grid rather than straight down the page. This view is
super cluttered and needs the support of whitespace, but is much
more palettable than a list. Perhaps a fixed number of pokes per
row is recommended.