use an effect to set a window scroll listener on component load
in order to know when to load new pokemons, the app needs to know
the position of the viewport relative to the content, allowing
for a generator call when the viewport is near or out of content.

It would be nice to monitor this in terms of request speed, and to
buffer earlier if requests take longer to fulfill.