provide a third variable to store the pokes that are requested
in order to retain next to satisfy the while clause, and to allow
for pre-emptive fetching of the pokemon list to reduce wait times
for loading the page, a third variable will be leveraged to store
the requested pokemon. Ensuring that there is no conflict between
curr.pokemon and next.pokemon, may require a different strategy in
the long term.